Move storybook into component folders
