+++ src/lib/components/Collapsible/stories/CollapsibleStyled.svelte
